matrix syntax fixes

+++ matrix/matrix.hpp
@@ -1,30 +1,29 @@
 #pragma once
 #include <vector>
 
-
 template <size_t N, size_t M, typename T = int64_t>
 class Matrix {
  public:
-  Matrix(): Matrix(T()) {};
+  Matrix() : Matrix(T()){};
 
   Matrix(const std::vector<std::vector<T>>& vector)
-    : matrix_(std::vector<std::vector<T>>(N, std::vector<T>(M))) {
-    for (int i = 0; i < N; ++i) {
-      for (int j = 0; j < M; ++j) {
+      : matrix_(std::vector<std::vector<T>>(N, std::vector<T>(M))) {
+    for (size_t i = 0; i < N; ++i) {
+      for (size_t j = 0; j < M; ++j) {
         matrix_[i][j] = vector[i][j];
       }
     }
   };
 
-  Matrix(const Matrix<N, M, T>& matrix): Matrix(matrix.matrix_) {};
+  Matrix(const Matrix<N, M, T>& matrix) : Matrix(matrix.matrix_){};
 
   Matrix(const T& elem) {
     matrix_ = std::vector<std::vector<T>>(N, std::vector<T>(M, elem));
   };
 
   Matrix& operator+=(const Matrix& other) {
-    for (int i = 0; i < N; ++i) {
-      for (int j = 0; j < M; ++j) {
+    for (size_t i = 0; i < N; ++i) {
+      for (size_t j = 0; j < M; ++j) {
         matrix_[i][j] += other(i, j);
       }
     }
@@ -32,40 +31,35 @@ class Matrix {
   };
 
   Matrix& operator-=(const Matrix& other) {
-    for (int i = 0; i < N; ++i) {
-      for (int j = 0; j < M; ++j) {
+    for (size_t i = 0; i < N; ++i) {
+      for (size_t j = 0; j < M; ++j) {
         matrix_[i][j] -= other(i, j);
       }
     }
     return *this;
   };
 
-
   Matrix& operator+=(T element) {
-    for (int i = 0; i < N; ++i) {
-      for (int j = 0; j < M; ++j) {
+    for (size_t i = 0; i < N; ++i) {
+      for (size_t j = 0; j < M; ++j) {
         matrix_[i][j] += element;
       }
     }
     return *this;
   };
 
-  Matrix& operator-=(T element) {
-    return (*this += (-element));
-  };
+  Matrix& operator-=(T element) { return (*this += (-element)); };
 
   Matrix& operator*=(T element) {
-    for (int i = 0; i < N; ++i) {
-      for (int j = 0; j < M; ++j) {
+    for (size_t i = 0; i < N; ++i) {
+      for (size_t j = 0; j < M; ++j) {
         matrix_[i][j] *= element;
       }
     }
     return *this;
   };
 
-  T& operator()(size_t row, size_t col) {
-    return matrix_[row][col];
-  };
+  T& operator()(size_t row, size_t col) { return matrix_[row][col]; };
 
   const T& operator()(size_t row, size_t col) const {
     return matrix_[row][col];
@@ -73,8 +67,8 @@ class Matrix {
 
   Matrix<M, N, T> Transposed() {
     Matrix<M, N, T> result;
-    for (int i = 0; i < N; ++i) {
-      for (int j = 0; j < M; ++j) {
+    for (size_t i = 0; i < N; ++i) {
+      for (size_t j = 0; j < M; ++j) {
         result(j, i) = matrix_[i][j];
       }
     }
@@ -84,7 +78,7 @@ class Matrix {
   T Trace() {
     static_assert(N == M);
     T result = 0;
-    for (int i = 0; i < N; ++i) {
+    for (size_t i = 0; i < N; ++i) {
       result += matrix_[i][i];
     }
     return result;
@@ -135,13 +129,12 @@ template <size_t N, size_t M, size_t K, typename T = int64_t>
 Matrix<N, K, T> operator*(const Matrix<N, M, T>& first,
                           const Matrix<M, K, T>& second) {
   Matrix<N, K, T> result;
-  for (int i = 0; i < N; ++i) {
-    for (int j = 0; j < K; ++j) {
-      for (int k = 0; k < M; ++k) {
+  for (size_t i = 0; i < N; ++i) {
+    for (size_t j = 0; j < K; ++j) {
+      for (size_t k = 0; k < M; ++k) {
         result(i, j) += first(i, k) * second(k, j);
       }
     }
   }
   return result;
 }
-
